Refactor authentication module
